+++ sync/cond_event.go
@@ -1,7 +1,5 @@
 // Copyright 2016 Aleksandr Demakin. All rights reserved.
 
-//+build !darwin
-
 package sync
 
 import (