+++ sync/event_sema.go
@@ -1,6 +1,6 @@
 // Copyright 2016 Aleksandr Demakin. All rights reserved.
 
-// +build freebsd linux
+// +build windows darwin
 
 package sync
 
@@ -16,8 +16,9 @@ import (
 )
 
 type event struct {
-	name   string
+	s      *Semaphore
 	region *mmf.MemoryRegion
+	name   string
 	lwe    *lwEvent
 }
 
@@ -30,11 +31,19 @@ func newEvent(name string, flag int, perm os.FileMode, initial bool) (*event, er
 	if err != nil {
 		return nil, errors.Wrap(err, "failed to create shared state")
 	}
-	state := allocator.ByteSliceData(region.Data())
+	s, err := NewSemaphore(name, flag, perm, 0)
+	if err != nil {
+		region.Close()
+		if created {
+			shm.DestroyMemoryObject(mutexSharedStateName(name, "s"))
+		}
+		return nil, errors.Wrap(err, "failed to create a semaphore")
+	}
 	result := &event{
-		lwe:    newLightweightEvent(state, &futex{ptr: state}),
+		lwe:    newLightweightEvent(allocator.ByteSliceData(region.Data()), newSemaWaiter(s)),
 		name:   name,
 		region: region,
+		s:      s,
 	}
 	if created {
 		result.lwe.init(initial)
@@ -55,20 +64,30 @@ func (e *event) waitTimeout(timeout time.Duration) bool {
 }
 
 func (e *event) close() error {
-	return e.region.Close()
+	e1, e2 := e.s.Close(), e.region.Close()
+	if e1 != nil {
+		return errors.Wrap(e1, "failed to close sema")
+	}
+	if e2 != nil {
+		return errors.Wrap(e2, "failed to close shared state")
+	}
+	return nil
 }
 
 func (e *event) destroy() error {
 	if err := e.close(); err != nil {
-		return errors.Wrap(err, "failed to close shm region")
+		return errors.Wrap(err, "failed to close the event")
 	}
 	return destroyEvent(e.name)
 }
 
 func destroyEvent(name string) error {
-	err := shm.DestroyMemoryObject(eventName(name))
-	if err != nil {
-		return errors.Wrap(err, "failed to destroy memory object")
+	e1, e2 := shm.DestroyMemoryObject(eventName(name)), destroySemaphore(name)
+	if e1 != nil {
+		return errors.Wrap(e1, "failed to destroy memory object")
+	}
+	if e2 != nil {
+		return errors.Wrap(e2, "failed to destroy semaphore")
 	}
 	return nil
 }