+++ sync/event_spin.go
@@ -1,6 +1,6 @@
 // Copyright 2016 Aleksandr Demakin. All rights reserved.
 
-// +build darwin
+// +build ignore
 
 package sync