+++ sync/lwmutex.go
@@ -14,27 +14,21 @@ const (
 	lwmStateSize = 4
 
 	lwmSpinCount         = 100
-	lwmUnlocked          = uint32(0)
-	lwmLockedNoWaiters   = uint32(1)
-	lwmLockedHaveWaiters = uint32(2)
+	lwmUnlocked          = int32(0)
+	lwmLockedNoWaiters   = int32(1)
+	lwmLockedHaveWaiters = int32(2)
 )
-
-// waitWaker is an object, which implements wake/wait semantics.
-type waitWaker interface {
-	wake(count uint32) (int, error)
-	wait(value uint32, timeout time.Duration) error
-}
 
 // lwMutex is a lightweight mutex implementation operating on a uint32 memory cell.
 // it tries to minimize amount of syscalls needed to do locking.
 // actual sleeping must be implemented by a waitWaker object.
 type lwMutex struct {
-	state *uint32
+	state *int32
 	ww    waitWaker
 }
 
 func newLightweightMutex(state unsafe.Pointer, ww waitWaker) *lwMutex {
-	return &lwMutex{state: (*uint32)(state), ww: ww}
+	return &lwMutex{state: (*int32)(state), ww: ww}
 }
 
 // init writes initial value into mutex's memory location.
@@ -49,7 +43,7 @@ func (lwm *lwMutex) lock() {
 }
 
 func (lwm *lwMutex) tryLock() bool {
-	return atomic.CompareAndSwapUint32(lwm.state, lwmUnlocked, lwmLockedNoWaiters)
+	return atomic.CompareAndSwapInt32(lwm.state, lwmUnlocked, lwmLockedNoWaiters)
 }
 
 func (lwm *lwMutex) lockTimeout(timeout time.Duration) bool {
@@ -69,33 +63,33 @@ func (lwm *lwMutex) doLock(timeout time.Duration) error {
 			return nil
 		}
 	}
-	old := atomic.LoadUint32(lwm.state)
+	old := atomic.LoadInt32(lwm.state)
 	if old != lwmLockedHaveWaiters {
-		old = atomic.SwapUint32(lwm.state, lwmLockedHaveWaiters)
+		old = atomic.SwapInt32(lwm.state, lwmLockedHaveWaiters)
 	}
 	for old != lwmUnlocked {
 		if err := lwm.ww.wait(lwmLockedHaveWaiters, timeout); err != nil {
 			return err
 		}
-		old = atomic.SwapUint32(lwm.state, lwmLockedHaveWaiters)
+		old = atomic.SwapInt32(lwm.state, lwmLockedHaveWaiters)
 	}
 	return nil
 }
 
 func (lwm *lwMutex) unlock() {
-	if old := atomic.LoadUint32(lwm.state); old == lwmLockedHaveWaiters {
+	if old := atomic.LoadInt32(lwm.state); old == lwmLockedHaveWaiters {
 		*lwm.state = lwmUnlocked
 	} else {
 		if old == lwmUnlocked {
 			panic("unlock of unlocked mutex")
 		}
-		if atomic.SwapUint32(lwm.state, lwmUnlocked) == lwmLockedNoWaiters {
+		if atomic.SwapInt32(lwm.state, lwmUnlocked) == lwmLockedNoWaiters {
 			return
 		}
 	}
 	for i := 0; i < lwmSpinCount; i++ {
 		if *lwm.state != lwmUnlocked {
-			if atomic.CompareAndSwapUint32(lwm.state, lwmLockedNoWaiters, lwmLockedHaveWaiters) {
+			if atomic.CompareAndSwapInt32(lwm.state, lwmLockedNoWaiters, lwmLockedHaveWaiters) {
 				return
 			}
 		}